refactor(api): split suspend C API into imp_api_suspend.cpp (file-size gate)
